+++ extensions/paymenter/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Paymenter Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- View Invoices
+- View Tickets
+- View & Create Users